[IMP] almik: fix required fields and add constrains on crm I#24600
An xpath was rewritting the required attribute on some fields, hence the
xpaths pointing to the same fields where unified to avoid rewritting and
missing attributes.
The same feature is requested by user on expected_revenue and
probability, since they are float fields a constrain is added to prevent
from leaving zeros by user.
Add translation for validation message.